feat：Refactoring page navigation
