+++ frontendv3/src/layouts/ManagerNav.vue
@@ -1,31 +1,20 @@
 <script setup lang="ts">
 import { ref, reactive, onMounted, inject, h } from 'vue';
-import { useRouter, RouterLink } from 'vue-router';
-import type { MenuOption } from "naive-ui";
-import { useMessage } from 'naive-ui';
+import { RouterLink } from 'vue-router';
 import { NIcon } from 'naive-ui';
-import type { Component } from 'vue';
+import type { MenuOption } from 'naive-ui';
 import {
   SettingsOutline as SettingIcon,
-  DocumentText as DocuIcon,
-  LogOutOutline as LogoutIcon,
-  PersonCircle as PersonIcon,
   AppsOutline as ProjectIcon,
   CloudOutline as CloudIcon,
   PeopleOutline as PeopleIcon,
   FolderOpenOutline as FolderIcon,
   CalendarOutline as CalendarIcon
 } from '@vicons/ionicons5';
-import UserApi from '~/request/user';
+import BaseNav from './BaseNav.vue';
 import ProjectApi from '~/request/project';
 import projectStorage from '~/store/index';
 
-// Declare the emitted events
-const emit = defineEmits<{
-  (e: 'changeThemeSignal'): void;
-}>();
-
-// Type definitions
 type TprojectOptions = {
   value: string;
   label: string;
@@ -33,36 +22,18 @@ type TprojectOptions = {
 
 type modelRef = {
   projectOptions: TprojectOptions[];
-  // envOptions: [];
 };
 
-// Reactive state
-const router = useRouter();
-const message = useMessage();
-const datas = reactive({
-  loading: false,
-  projectValue: 0,
-});
-const mainhtml = document.getElementsByTagName('html');
-const btnLabel = ref('深色');
-const localStorage = window.localStorage;
-const model = ref<modelRef>({
-  projectOptions: [],
-  // envOptions: [],
-});
-const token = ref<string | null>('');
+const renderIcon = (icon: any) => {
+  return () => h(NIcon, null, { default: () => h(icon) });
+};
 
-// Injected function
-const reload = inject<() => void>('reload');
-
-// Menu options
-const renderIcon = (icon: Component) => () => h(NIcon, null, { default: () => h(icon) });
-
-const menuOptions = [
+const menuOptions: MenuOption[] = [
   {
     label: () =>
       h(
-        RouterLink, {
+        RouterLink,
+        {
           to: {
             name: 'manager-Case',
             params: {
@@ -78,7 +49,8 @@ const menuOptions = [
   {
     label: () =>
       h(
-        RouterLink, {
+        RouterLink,
+        {
           to: {
             name: 'manager-Task',
             params: {
@@ -99,7 +71,8 @@ const menuOptions = [
       {
         label: () =>
           h(
-            RouterLink, {
+            RouterLink,
+            {
               to: {
                 name: 'center-Project',
                 params: {
@@ -115,8 +88,9 @@ const menuOptions = [
       {
         label: () =>
           h(
-            RouterLink, {
-              to: { 
+            RouterLink,
+            {
+              to: {
                 name: 'center-Env',
                 params: {
                   lang: 'zh-CN'
@@ -131,8 +105,9 @@ const menuOptions = [
       {
         label: () =>
           h(
-            RouterLink, {
-              to: { 
+            RouterLink,
+            {
+              to: {
                 name: 'center-Team',
                 params: {
                   lang: 'zh-CN'
@@ -148,21 +123,17 @@ const menuOptions = [
   },
 ];
 
-// 个人选项
-const  personOptions = [
-    {
-      label: '操作手册',
-      key: 'help_documentation',
-      icon: renderIcon(DocuIcon),
-    },
-    {
-      label: '退出登录',
-      key: 'logout',
-      icon: renderIcon(LogoutIcon),
-    },
-]
+const datas = reactive({
+  loading: false,
+  projectValue: 0,
+});
 
-// Methods
+const model = ref<modelRef>({
+  projectOptions: [],
+});
+
+const reload = inject<() => void>('reload');
+
 const initProjectList = async () => {
   datas.loading = true;
   const resp = await ProjectApi.getProjects();
@@ -171,97 +142,29 @@ const initProjectList = async () => {
       value: project.id,
       label: project.name,
     }));
-  } else {
-    message.error(resp.error.message); 
   }
   datas.loading = false;
 };
 
 const changeProject = (value: number, option: any) => {
-  // 保存项目缓存
-  projectStorage.setProject(value, option.label)
-
+  projectStorage.setProject(value, option.label);
   reload?.();
 };
 
-const changeTheme = () => {
-  if (btnLabel.value === '深色') {
-    btnLabel.value = '浅色';
-    localStorage.setItem('themeMode', 'dark');
-  } else {
-    btnLabel.value = '深色';
-    localStorage.setItem('themeMode', 'light');
-  }
-  emit('changeThemeSignal');
-  mainhtml[0].classList.toggle('dark');
-};
-
-const handleSelect = (key: string | number) => {
-  switch (key) {
-    case 'logout':
-      UserApi.logout({ token: token.value }).then((resp: any) => {
-        if (resp.success) {
-          sessionStorage.clear();
-          router.push('/login');
-        } else {
-          message.error(resp.error.message);
-        }
-      });
-      break;
-    case 'help_documentation':
-      window.open('https://github.com/SeldomQA/seldom-platform', '_blank');
-      break;
-    default:
-      break;
-  }
-};
-
-// New method: handleUpdateValue
-const handleUpdateValue = (key: string, item: MenuOption) => {
-  console.log('[onUpdate] :key=' + JSON.stringify(key) +' item = '+ JSON.stringify(item));
-};
-
-// Lifecycle hooks
 onMounted(() => {
-  const mode = localStorage.getItem('themeMode');
-  btnLabel.value = mode === 'light' || mode == null ? '深色' : '浅色';
-  token.value = sessionStorage.getItem('token');
-
   initProjectList();
-
-   const projectData = projectStorage.getProject()
+  const projectData = projectStorage.getProject();
   if (projectData) {
-    datas.projectValue = projectData?.id ?? 0
+    datas.projectValue = projectData?.id ?? 0;
   }
-});
-
-// Expose to template
-defineExpose({
-  menuOptions,
-  datas,
-  btnLabel,
-  changeTheme,
-  model,
-  changeProject,
-  handleSelect,
-  handleUpdateValue,
-  token,
-  PersonIcon,
-  personOptions
 });
 </script>
 
 <template>
-  <div class="header">
-    <n-space justify="space-between">
-      <div>
-        <span style="float: left;">
-          <img src="../assets/seldom-platform.gif" style="height: 40px" />
-        </span>
-        <span style="float: right;">
-          <n-tag type="warning"> 项目管理 </n-tag>
-        </span>
-      </div>
+  <BaseNav 
+    :menu-options="menuOptions"
+    :tag="{ type: 'warning', text: '项目管理' }">
+    <template #project-select>
       <n-form inline :model="model" label-placement="left">
         <n-form-item label="">
           <n-select
@@ -270,27 +173,11 @@ defineExpose({
             placeholder="选择项目"
             @update:value="changeProject"
             v-model:value="datas.projectValue"
-          >
-          </n-select>
+          />
         </n-form-item>
       </n-form>
-      <n-menu :options="menuOptions"  @update:value="handleUpdateValue" mode="horizontal"/>
-      <n-space>
-        <n-button @click="changeTheme">
-          <template #default>
-            {{ btnLabel }}
-          </template>
-        </n-button>
-        <n-dropdown :options="personOptions" @select="handleSelect">
-          <n-button>
-            <template #icon>
-              <n-icon :component="PersonIcon"></n-icon>
-            </template>
-          </n-button>
-        </n-dropdown>
-      </n-space>
-    </n-space>
-  </div>
+    </template>
+  </BaseNav>
 </template>
 
 <style scoped>
