--- frontendv3/src/layouts/footer.vue
@@ -1,15 +1,0 @@
-<template>
-  <div>
-    <n-layout-footer bordered position="absolute" style="text-align: center">
-        © 2024 Powered by SeldomQA Team
-    </n-layout-footer>
-  </div>
-</template>
-
-<style scoped>
-.n-layout-header,
-.n-layout-footer {
-  /* background: rgba(128, 128, 128, 0.2); */
-  padding: 20px;
-}
-</style>
