+++ frontendv3/src/pages/Center.vue
@@ -1,60 +1,8 @@
 <script setup lang="ts">
+import BaseLayout from "~/layouts/BaseLayout.vue";
 import CenterNav from "~/layouts/CenterNav.vue";
-import Footer from "~/layouts/footer.vue";
-import { darkTheme } from "naive-ui";
-import type { GlobalTheme } from "naive-ui";
-import { ref, onMounted, watch, provide, nextTick } from "vue";
-
-const theme = ref<GlobalTheme | null>(null);
-const isRouterAlive = ref(true);
-
-const changeTheme = () => {
-  theme.value = theme.value ? null : darkTheme;
-  const mode = theme.value ? "dark" : "light";
-  localStorage.setItem("themeMode", mode);  // 存储模式
-  document.documentElement.classList.toggle("dark", mode === "dark");
-};
-
-const reload = () => {
-  isRouterAlive.value = !isRouterAlive.value;
-};
-
-provide("reload", reload);
-
-onMounted(() => {
-  const mode = localStorage.getItem("themeMode") ?? "light";  // 默认 "light"
-  theme.value = mode === "dark" ? darkTheme : null;  // 根据 localStorage 设置主题
-  document.documentElement.classList.toggle("dark", mode === "dark");
-});
 </script>
 
 <template>
-  <n-config-provider :theme="theme">
-  <n-layout class="dflayout" has-sider position="absolute">
-    <n-layout>
-      <n-layout-header bordered>
-        <CenterNav @changeThemeSignal="changeTheme"></CenterNav>
-      </n-layout-header>
-      <n-layout 
-        position="absolute" 
-        :style="theme === null ? 'top: 76px; bottom: 64px; background-color: #fbf9f6': 'top: 76px; bottom: 64px;' ">
-        <n-layout 
-        :content-style="theme === null ? 'padding: 24px; background-color: #fbf9f6' : 'padding: 24px'">
-          <div v-if="isRouterAlive">
-            <router-view></router-view>
-          </div>
-        </n-layout>
-      </n-layout>
-      <Footer></Footer>
-    </n-layout>
-  </n-layout>
-</n-config-provider>
+  <BaseLayout :nav-component="CenterNav" />
 </template>
-
-<style scoped>
-.n-layout-header,
-.n-layout-footer {
-  /* background: rgba(128, 128, 128, 0.2); */
-  padding: 20px;
-}
-</style>
